put randman classes in separate files
